requirements.txt
requests
sseclient-py
python-telegram-bot
tavily-python